third_party/blink/web_tests/platform/mac-mac13/external/wpt/visual-viewport/viewport-scrollbars-cause-resize-in-iframe-expected.txt

This is a testharness.js-based test.
[FAIL] the appearance of classic scrollbars will fire a resize eventat window.visualViewport in iframe
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
